Rename deprecated button modules
