--- src/components/button/index.ts
@@ -1,5 +1,0 @@
-import { Button } from './button'
-import type { ButtonProps, ButtonSize, ButtonVariant } from './button'
-
-export type { ButtonProps, ButtonSize, ButtonVariant }
-export default Button
