+++ src/components/deprecated-button/deprecated-button.less
