+++ src/components/deprecated-button/index.ts
@@ -1,0 +1,5 @@
+import { Button } from './deprecated-button'
+import type { ButtonProps, ButtonSize, ButtonVariant } from './deprecated-button'
+
+export type { ButtonProps, ButtonSize, ButtonVariant }
+export default Button
