+++ stories/components/Tooltip.stories.tsx
@@ -3,8 +3,8 @@ import './styles/tooltip_story.less'
 import React from 'react'
 
 import { Tooltip, TooltipProps } from '../../src/components/tooltip'
-import Button from '../../src/components/button'
-import type { ButtonProps } from '../../src/components/button'
+import Button from '../../src/components/deprecated-button'
+import type { ButtonProps } from '../../src/components/deprecated-button'
 
 // Story setup ================================================================
